# docs/openapi.yaml
openapi: 3.1.0
info:
  title: NetVigil API
  version: 0.1.0
  description: |
    NetVigil is an AI-driven Network Detection and Response platform for SMBs.
    This document is the single source of truth for the contract between the
    FastAPI backend (`services/api`) and the TypeScript clients
    (`packages/web`, `packages/mobile`). Generated TypeScript types live in
    `packages/shared-types/src/openapi.d.ts` and must not be edited by hand.

    Authentication: bearer JWT (RS256), 15-minute access tokens with refresh
    rotation. All tenant-scoped endpoints require the bearer token.
  contact:
    name: NIT3003 Capstone Team
  license:
    name: Educational use — Victoria University
servers:
  - url: http://localhost:8000/api/v1
    description: Local dev (docker-compose)
  - url: https://staging.netvigil.example/api/v1
    description: Staging (AWS ECS)
...
paths:
...
  /auth/register:
    post:
      tags: [auth]
      summary: Register a new organisation and admin user
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RegisterRequest' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '409':
          description: Email already registered
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /auth/login:
    post:
      tags: [auth]
      summary: Exchange credentials for access + refresh tokens
      description: |
        Returns 401 with a generic message on any auth failure. Never
        reveals whether the user exists.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LoginRequest' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /auth/refresh:
    post:
      tags: [auth]
      summary: Rotate refresh token, return new access token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /auth/me:
    get:
      tags: [auth]
      summary: The authenticated user's profile
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    BadRequest:
      description: Validation failure
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Unauthorized:
      description: Invalid or missing credentials
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code: { type: string, example: invalid_credentials }
        message: { type: string }
        details:
          type: object
          additionalProperties: true

    # ─── Auth ────────────────────────────────────────────────────────────
    RegisterRequest:
      type: object
      required: [email, password]
      description: |
        Exactly one of organizationName (create new org) or organizationId (join existing) must be set.
        Creating a new org → role is forced to admin, status active.
        Joining an existing org → chosen role, status pending (awaiting approval).
      properties:
        organizationName:
          type: string
          minLength: 2
          maxLength: 200
          nullable: true
        organizationId:
          type: string
          format: uuid
          nullable: true
        email: { type: string, format: email }
        password: { type: string, minLength: 12 }
        role:
          $ref: '#/components/schemas/UserRole'
        timezone: { type: string, example: Australia/Brisbane }
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string }
    AuthResponse:
      type: object
      properties:
        accessToken: { type: string }
        refreshToken: { type: string }
        expiresIn:
          type: integer
          description: Seconds until access token expires
          example: 900
        user:
          $ref: '#/components/schemas/User'
          nullable: true
        mfaRequired:
          type: boolean
          default: false
        mfaToken:
          type: string
          nullable: true
    UserRole:
      type: string
      enum:
        - super_admin
        - admin
        - senior_analyst
        - analyst
        - threat_hunter
        - forensic_investigator
        - auditor
        - developer
    UserStatus:
      type: string
      enum: [pending, active, rejected]
    User:
      type: object
      required: [id, organizationId, email, role, status, mfaEnrolled, createdAt]
      properties:
        id: { type: string, format: uuid }
        organizationId: { type: string, format: uuid }
        email: { type: string, format: email }
        role: { $ref: '#/components/schemas/UserRole' }
        status: { $ref: '#/components/schemas/UserStatus' }
        mfaEnrolled: { type: boolean }
        createdAt: { type: string, format: date-time }
...
security:
  - bearerAuth: []